requirements.txt
# Python Requirements for IDO Compiler & Decompiler
# Install with: pip install -r requirements.txt

# DDS Converter dependencies
Pillow>=10.0.0

# Note: IDO tool (ido_tool.py) uses only Python standard library:
# - sys
# - os  
# - zlib
# - struct
# - json
# - pathlib
# - csv

# All required modules are built into Python 3.x
